Lab Nine Update
Completed on March 30, 2022
Lab nine was an introduction to using the Texas Instruments MSP430FR2433. We used CCS to compile and run the code created here. 
The code we were given turned on LED 1 and flashed it on and off, from there we changed the code to make LED 2 flash rather than LED 1. We then made it flash faster. Lastly, we demonstrated going through the while loop one step at a time, which is why it is set to 5 instead of a smaller value.