Lab 10 Update
For this lab, we created a "Multi-Function Flashlight" using the same board as lab nine, with an additional component that was provided that has buttons and a few different LED colors. 

The LEDs start off. One button press would give us a yellow LED turning on, a second button press would give us both red LEDs turning on (with the yellow turning off), a third button press would give us a green LED turns on (with the red turning off), a fourth would make the green LED flash (rather than it's previous solid state), and a fifth press would bring us back to the start.